12/23 반제품 - 커핑 UI화면 셋팅

+++ Coffee_iOS/Coffee_iOS/VC/HomeVC.h
@@ -10,7 +10,7 @@
 
 @interface HomeVC : UIViewController{
     NSMutableArray *tableList;
-    int tab_position;
+    NSInteger tab_position;
 }
 
 @property (weak, nonatomic) IBOutlet UITableView *homeTableView;
